dummy

+++ cicd/build/jenkins_build_rpm.sh
@@ -16,6 +16,10 @@ echo "(DEBUG)   \- PAYLOAD: $PAYLOAD"
 # example of PAYLOAD: https://dmapelli.web.cern.ch/public/crab/20220127/crabserver_github_release_payload_example.json
 echo "(DEBUG) end"
 
+if [[ -z $RELEASE_TAG ]]; then echo '$RELEASE_TAG' "is empty, exiting"; exit 1 fi
+if [[ -z $CRABSERVER_REPO ]]; then echo '$CRABSERVER_REPO' "is empty, exiting"; exit 1 fi
+if [[ -z $WMCORE_REPO ]]; then echo '$WMCORE_REPO' "is empty, exiting"; exit 1 fi
+
 #do a clean up
 docker system prune -af
 
@@ -45,6 +49,7 @@ else
    export BRANCH=$(git branch -a --contains tags/$RELEASE_TAG | sed "s/*//g" | sed "s/ //g" | grep origin | awk -F "/" '{print $3}')
    cd ..
 fi
+if [[ -z $BRANCH ]]; then echo '$BRANCH' "is empty, exiting"; exit 1 fi
 echo "BRANCH=${BRANCH}" >> $WORKSPACE/properties_file
 # ACHTUNG!
 # $BRANCH can contain only alfanumeric characters and the underscore "_". for example, the dash "-" is not allowed
@@ -54,17 +59,12 @@ echo "BRANCH=${BRANCH}" >> $WORKSPACE/properties_file
 if [ ${WMCORE_REPO} == "dmwm" ]; then
    WMCORE_TAG=$(grep -oP "wmcver==\K.*" CRABServer/requirements.txt)
 fi
+if [[ -z $WMCORE_TAG ]]; then echo '$WMCORE_TAG' "is empty, exiting"; exit 1 fi
+
 
 echo "(DEBUG) env variables that could have been updated from the default:"
 echo "(DEBUG)   \- BRANCH (crabserver): $BRANCH"
 echo "(DEBUG)   \- WMCORE_TAG: $WMCORE_TAG"
-
-# eventually: exit if the input variables are not set.
-# exit if one of the required variables is not set
-if [[ -z $BRANCH ]]; then
-  echo '$BRANCH' "is empty, exiting"
-  exit 1 
-fi
 
 ##### All the required env variables are set at this point. Do not set any input below this line
 #update .spec files with new CRABServer and  WMCore tags; update from which branch RPMs should be built
